+++ README.md
@@ -1,11 +1,26 @@
-| テーブル名 | カラム名　|  データ型 |
-|--------| --------| ---------|
-|User|name|string|
-|    |email|string|
-|    |password|string|
-|Task|task_name|string|
-|    |task_content|text|
-|    |deadline|string|
-|    |status|string|
-|Label|label_name|string|
-|     |label_content|string|
+
+| モデル名 | カラム名 | データ型 |
+|----------|----------|----------|
+| User     | id       | integer  |
+|          | name     | string   |
+|          | email    | string   |
+
+| モデル名 | カラム名     | データ型 |
+|----------|--------------|----------|
+| Task     | id           | integer  |
+|          | task_name    | string   |
+|          | task_content | string   |
+|          | deadline     | string   |
+|          | status       | string   |
+|          | priority     | string   |
+
+
+| モデル名 | カラム名   | データ型 |
+|----------|------------|----------|
+| Label    | id         | integer  |
+|          | label_name | string   |
+
+| モデル名   | カラム名 | データ型   |
+|------------|----------|------------|
+| Task_Label | task_id  | references |
+|            | label_id | references |
