+++ app/assets/stylesheets/tasks.scss
@@ -1,3 +1,3 @@
-// Place all the styles related to the Task controller here.
+// Place all the styles related to the tasks controller here.
 // They will automatically be included in application.css.
 // You can use Sass (SCSS) here: https://sass-lang.com/
