+++ app/controllers/tasks_controller.rb
@@ -1,0 +1,59 @@
+class TasksController < ApplicationController
+    before_action :set_task, only: [:show, :edit, :update, :destroy]
+
+    def index
+      @tasks = Task.all
+    end
+
+    def new
+        @task = Task.new
+    end
+
+    def confirm
+      #@task = current_user.tasks.build(task_params)
+      render :new if @task.invalid?
+    end
+
+    def create
+      @task = Task.new(task_params)
+      if @task.save
+          flash[:notice] = "作成しました！"
+          redirect_to tasks_path
+      else
+          render :new
+      end
+    end
+
+    def edit
+      #@task = task.find(params[:id])
+    end
+
+
+    def show
+      #@task = task.find(params[:id])
+    end
+
+
+    def update
+      if @task.update(task_params)
+        redirect_to tasks_path, notice: "編集しました！"
+      else
+        render :edit
+      end
+    end
+
+    def destroy
+      @task.destroy
+      redirect_to tasks_path, notice:"削除しました！"
+    end
+
+  private
+
+    def task_params
+      params.require(:task).permit(:task_content, :task_name)
+    end
+
+    def set_task
+      @task = Task.find(params[:id])
+    end
+end
