--- app/helpers/task_helper.rb
@@ -1,2 +1,0 @@
-module TaskHelper
-end
