+++ app/helpers/tasks_helper.rb
@@ -1,0 +1,2 @@
+module TasksHelper
+end
