--- app/views/task/_form.html.erb
@@ -1,29 +1,0 @@
-  <% if @task.errors.any? %>
-    <div id="error_explanation">
-      <h2><%= @task.errors.count %>件のエラーがあります。</h2>
-      <ul>
-
-      <% @task.errors.full_messages.each do |msg| %>
-        <li><%= msg %></li>
-      <% end %>
-      </ul>
-    </div>
-  <% end %>
-<%= form_with(model: @task, local: true ,url: choose_new_or_edit ) do |form| %>
-  <div class="task_title">
-    <%= form.label :title %>
-    <%= form.text_field :title %>
-  </div>
-  <div class="task_title">
-    <%= form.label :content %>
-    <%= form.text_field :content %>
-  </div>
-  <div class="field">
-    <%= form.label :image %>
-    <%= form.file_field :image %>
-    <%= form.hidden_field :image_cache %>
-  </div>
-    <%= form.submit "送信"%>
-<% end %>
-
-<%= link_to "一覧画面にもどる", tasks_path %>
