--- app/views/task/confirm.html.erb
@@ -1,13 +1,0 @@
-<h3>以下の内容で、送信する。</h3>
-
-<%= form_with(model: @task, local: true) do |form| %>
-  <%= form.hidden_field :title %>
-  <%= form.hidden_field :content %>
-
-  <p>タイトル:<%= @task.title %></p>
-  <p>本文:<%= @task.content %></p>
-  <%= image_tag(@task.image.url) if @task.image && @task.image.url %>
-  <%= form.hidden_field :image_cache %>
-  <%= form.submit  "登録する" %>
-  <%= form.submit "戻る", name: 'back' %>
-<% end %>
