+++ app/views/tasks/index.html.erb
@@ -6,9 +6,9 @@
   <tr>
     <th>一覧<th>
   </tr>
-<% @task.each do |task| %>
+<% @tasks.each do |task| %>
   <tr>
-    <td><%= task.content %></td>
+    <td><%= task.task_content %></td>
     <td><%= link_to '確認', task_path(task.id), class: "link" %></td>
     <td><%= link_to '編集', edit_task_path(task.id), class: "link" %></td>
     <td><%= link_to '削除', task_path(task.id), method: :delete, data: { confirm: '本当に削除していいですか？'}, class: "link" %></td>
@@ -16,6 +16,5 @@
 <% end%>
 </table>
 
-<div class="new_button">
 <%= link_to'新しく投稿する', new_task_path, class: "new_link" %>
-</div>
+
