+++ app/views/tasks/show.html.erb
@@ -1,0 +1,9 @@
+<header>
+  <h2>詳細画面</h2>
+</header>
+
+<p>名前:<%= @task.task_name %></p>
+<p>内容<%= @task.task_content %></p>
+
+
+<%= link_to "一覧画面にもどる", tasks_path%>
