+++ db/migrate/20230111023013_change_column_null_tasks.rb
@@ -1,0 +1,6 @@
+class ChangeColumnNullTasks < ActiveRecord::Migration[6.0]
+  def change
+    change_column_null :tasks, :task_name, false
+    change_column_null :tasks, :task_content, false
+  end
+end
