+++ test/controllers/tasks_controller_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class TaskControllerTest < ActionDispatch::IntegrationTest
+class TasksControllerTest < ActionDispatch::IntegrationTest
   # test "the truth" do
   #   assert true
   # end
